# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WebsiteLeadInputNiche = typeof WebsiteLeadInputNiche[keyof typeof WebsiteLeadInputNiche];


export const WebsiteLeadInputNiche = {
  Portfolio: 'Portfolio',
  Restaurant: 'Restaurant',
  Clinic: 'Clinic',
  'Salon_&_Beauty': 'Salon & Beauty',
  Fashion: 'Fashion',
  Real_Estate: 'Real Estate',
  Education: 'Education',
  Other: 'Other',
} as const;

export interface WebsiteLeadInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  businessName: string;
  /**
     * @maxLength 254
     * @nullable
     */
  email: string | null;
  /**
     * @minLength 7
     * @maxLength 32
     * @nullable
     */
  phone: string | null;
  niche: WebsiteLeadInputNiche;
}

export type WebsiteLeadNiche = typeof WebsiteLeadNiche[keyof typeof WebsiteLeadNiche];


export const WebsiteLeadNiche = {
  Portfolio: 'Portfolio',
  Restaurant: 'Restaurant',
  Clinic: 'Clinic',
  'Salon_&_Beauty': 'Salon & Beauty',
  Fashion: 'Fashion',
  Real_Estate: 'Real Estate',
  Education: 'Education',
  Other: 'Other',
} as const;

export interface WebsiteLead {
  id: number;
  name: string;
  /** @nullable */
  businessName: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phone: string | null;
  niche: WebsiteLeadNiche;
  createdAt: string;
}

export interface WebsiteLeadNicheCount {
  niche: string;
  count: number;
}

export interface WebsiteLeadSummary {
  total: number;
  byNiche: WebsiteLeadNicheCount[];
}

export interface ApiError {
  error: string;
}