# lib/api-zod/src/generated/types/websiteLead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WebsiteLeadNiche } from './websiteLeadNiche';

export interface WebsiteLead {
  id: number;
  name: string;
  /** @nullable */
  businessName: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  phone: string | null;
  niche: WebsiteLeadNiche;
  createdAt: Date;
}