# lib/api-zod/src/generated/types/websiteLeadInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WebsiteLeadInputNiche } from './websiteLeadInputNiche';

export interface WebsiteLeadInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  businessName: string;
  /**
     * @maxLength 254
     * @nullable
     */
  email: string | null;
  /**
     * @minLength 7
     * @maxLength 32
     * @nullable
     */
  phone: string | null;
  niche: WebsiteLeadInputNiche;
}